Implemented pagetoken and dropped unsetting
Thanks to @bristweb for this update! Seems I failed to properly implement the pagetoken and there was a block of code where I was unsetting parameters because at the time it seemed as if Google was not ignoring them as they claimed they would (a/k/a they "lied").

I'm not overly fond of the `sleep(3)` but it's a necessary evil because Google's pagetokens are not immediately available (as per https://developers.google.com/places/documentation/search#PlaceSearchPaging). My thought would be to perhaps create a loop that will retry X times every Y seconds to better compensate for inconsistent activation times on Google's end.

Closes #4 (for the most part)